Added yearly Transactions in the viewModel

+++ app/src/main/java/com/example/expensetracker/data/entities/Account.kt
@@ -9,5 +9,5 @@ data class Account(
     val accountId: String,
     val accountName: String?,
     val defaultCategory: String?,
-    var totalTransaction: Int
+    var totalTransaction: Double
 )
